Added card.js for better code readibility in main.js
