+++ src/frontend/src/main.js
@@ -8,8 +8,8 @@ import { matchEndScene } from "./components/winandloss";
 import CameraControls from "camera-controls";
 import { createCharacters, GameState } from "./components/gameState";
 import { characterDeath } from "./components/chardeath";
-import { logInfo, LogEventTypes } from "./components/logger";
-import { workingToColorSpace } from "three/webgpu";
+import { logInfo } from "./components/logger";
+import { createCard, cardCounts } from "./components/cards";
 
 const scene = new THREE.Scene();
 const camera = createCamera();
@@ -36,112 +36,36 @@ async function setupGame() {
   CameraControls.install({ THREE });
   const cameraControls = new CameraControls(camera, renderer.domElement);
   setControls(cameraControls);
-  
+  const characters = document.querySelector("#characters");
+  createCard(characters, 0, "https://placecats.com/100/100", "Cat", [
+    "square",
+    "square",
+  ]);
+  createCard(characters, 1, "https://placecats.com/100/100", "Cat", [
+    "square",
+    "square",
+  ]);
+  createCard(characters, 2, "https://placecats.com/100/100", "Cat", [
+    "square",
+    "square",
+  ]);
+  createCard(characters, 3, "https://placecats.com/100/100", "Cat", [
+    "square",
+    "square",
+    "square",
+  ]);
+  createCard(characters, 4, "https://placecats.com/100/100", "Cat", [
+    "square",
+    "square",
+  ]);
+  characterDeath(document.querySelector('[char-id="2"]'));
   render(cameraControls, spotlightHelper);
+  const logger = document.querySelector(".logs");
+  logInfo("Hello this is a fucking cat.");
+  matchEndScene("loss");
 }
 
 setupGame().catch(console.error);
-
-
-
-
-const cardCounts = {};
-
-function initializeCardCounts(charId, tokenTypes) {
-  if (!cardCounts[charId]) {
-    cardCounts[charId] = {};
-  }
-  tokenTypes.forEach((_, index) => {
-    cardCounts[charId][index + 1] = 0;
-  });
-}
-
-function updateTokenCount(token) {
-  //This is the token clicking hook here you can change and get info of the selected token ,inc or dec the count corresponding count to it.
-  //Remember to update the cardCounts data structure accordingly for smoother data storing.
-  const character = token.closest("#character-block");
-  const characterId = character.getAttribute("char-id");
-  const p = token.nextElementSibling;
-  const tokenId = token.getAttribute("token-no");
-  cardCounts[characterId][tokenId] += 1;
-  console.log(`The new value is ${cardCounts[characterId][tokenId]}`);
-  console.log("Decreament here and add changes according to yourself.");
-  p.textContent = `x${cardCounts[characterId][tokenId]}`;
-}
-
-function tooltip(tokenType) {
-  if (tokenType === "square") return `This is a trap.`;
-  else return `This is nothing.`;
-}
-
-function createCard(parent, charId, imgSrc, imgAlt, tokenTypes = []) {
-  initializeCardCounts(charId, tokenTypes);
-  const characterName = imgAlt;
-  const characterBlock = document.createElement("div");
-  characterBlock.id = "character-block";
-  characterBlock.setAttribute("char-id", charId);
-  const characterImgContainer = document.createElement("div");
-  characterImgContainer.className = "character-img-container";
-  const characterImg = document.createElement("img");
-  characterImg.className = "character-img";
-  characterImg.src = imgSrc;
-  characterImg.alt = imgAlt;
-  characterImgContainer.appendChild(characterImg);
-  const cardInfo = document.createElement("div");
-  cardInfo.className = "card-info";
-  const cardInfoP = document.createElement("p");
-  cardInfoP.className = "card-info-p";
-  cardInfoP.textContent = `My name is ${characterName}`;
-  const cardInfoToken = document.createElement("div");
-  cardInfoToken.className = "card-info-token";
-  tokenTypes.forEach((tokenType, index) => {
-    const token = document.createElement("div");
-    token.classList.add("token", tokenType);
-    token.setAttribute("data-tooltip", `${tooltip(tokenType)}`);
-    token.setAttribute("token-no", index + 1);
-    token.addEventListener("click", () => {
-      console.log(
-        `Clicked ${tokenType} card for Character ${charId}, Token No: ${
-          index + 1
-        }`
-      );
-      updateTokenCount(token);
-    });
-    const cardCount = document.createElement("p");
-    cardCount.classList.add("card-count");
-    cardCount.textContent = "x0";
-    cardInfoToken.appendChild(token);
-    cardInfoToken.appendChild(cardCount);
-  });
-  cardInfo.appendChild(cardInfoP);
-  cardInfo.appendChild(cardInfoToken);
-  characterBlock.appendChild(characterImgContainer);
-  characterBlock.appendChild(cardInfo);
-  parent.appendChild(characterBlock);
-}
-
-const characters = document.querySelector("#characters");
-createCard(characters, 0, "https://placecats.com/100/100", "Cat", [
-  "square",
-  "square",
-]);
-createCard(characters, 1, "https://placecats.com/100/100", "Cat", [
-  "square",
-  "square",
-]);
-createCard(characters, 2, "https://placecats.com/100/100", "Cat", [
-  "square",
-  "square",
-]);
-createCard(characters, 3, "https://placecats.com/100/100", "Cat", [
-  "square",
-  "square",
-  "square",
-]);
-createCard(characters, 4, "https://placecats.com/100/100", "Cat", [
-  "square",
-  "square",
-]);
 
 window.addEventListener(
   "resize",
@@ -155,8 +79,3 @@ window.addEventListener(
   },
   false
 );
-
-characterDeath(document.querySelector('[char-id="2"]'));
-const logger = document.querySelector(".logs");
-logInfo("Hello this is a fucking cat.");
-matchEndScene("loss");
